remove code duplications
