+++ src/Core/Types.Filters/Expressions/Comparable/ComparableGreaterThanOrEqualsOperationHandler.cs
@@ -1,59 +1,34 @@
-using System;
 using System.Linq.Expressions;
-using HotChocolate.Language;
-using HotChocolate.Utilities;
 
 namespace HotChocolate.Types.Filters.Expressions
 {
-    public class ComparableGreaterThanOrEqualsOperationHandler
-        : IExpressionOperationHandler
+    public sealed class ComparableGreaterThanOrEqualsOperationHandler
+        : ComparableOperationHandlerBase
     {
-        public bool TryHandle(
+        protected override bool TryCreateExpression(
             FilterOperation operation,
-            IInputType type,
-            IValueNode value,
-            Expression instance,
-            ITypeConversion converter,
+            MemberExpression property,
+            object parsedValue,
             out Expression expression)
         {
-            if (operation.Type == typeof(IComparable)
-                && (value is IntValueNode
-                    || value is FloatValueNode
-                    || value is EnumValueNode
-                    || value.IsNull()))
+            switch (operation.Kind)
             {
-                MemberExpression property =
-                    Expression.Property(instance, operation.Property);
-                var parsedValue = type.ParseLiteral(value);
+                case FilterOperationKind.GreaterThanOrEquals:
+                    expression = FilterExpressionBuilder.GreaterThanOrEqual(
+                        property, parsedValue);
+                    return true;
 
-                if (operation.Property.PropertyType
-                    .IsInstanceOfType(parsedValue))
-                {
-                    parsedValue = converter.Convert(
-                        typeof(object),
-                        operation.Property.PropertyType,
-                        parsedValue);
-                }
+                case FilterOperationKind.NotGreaterThanOrEquals:
+                    expression = FilterExpressionBuilder.Not(
+                        FilterExpressionBuilder.GreaterThanOrEqual(
+                            property, parsedValue)
+                     );
+                    return true;
 
-                switch (operation.Kind)
-                {
-
-                    case FilterOperationKind.GreaterThanOrEquals:
-                        expression = FilterExpressionBuilder.GreaterThanOrEqual(
-                            property, parsedValue);
-                        return true;
-
-                    case FilterOperationKind.NotGreaterThanOrEquals:
-                        expression = FilterExpressionBuilder.Not(
-                            FilterExpressionBuilder.GreaterThanOrEqual(
-                                property, parsedValue)
-                         );
-                        return true;
-                }
+                default:
+                    expression = null;
+                    return false;
             }
-
-            expression = null;
-            return false;
         }
     }
 }
